sort vars according to datasize to pack then tightly

+++ contracts/lib/Structs.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.0;
 library Structs {
     struct Vote {
         uint32 epoch;
-        uint32[] values;
+        uint48[] values;
     }
 
     struct Commitment {
